Revert "[supervisor] improve docker activation (#18284)"
This reverts commit 9bb400b0e8e419002850a70cd1274122a7fc59a1.
