+++ components/supervisor/pkg/activation/activation.go
@@ -10,8 +10,7 @@ import (
 	"os"
 	"sync"
 
-	"github.com/gitpod-io/gitpod/common-go/log"
-	"github.com/moredure/easygo/netpoll"
+	"github.com/mailru/easygo/netpoll"
 	"golang.org/x/xerrors"
 )
 
@@ -20,22 +19,16 @@ type Callback func(socketFD *os.File) error
 
 // Listen polls on the listener and calls callback when someone writes to it.
 func Listen(ctx context.Context, l net.Listener, activate Callback) error {
-	poller, err := netpoll.New(&netpoll.Config{
-		OnWaitError: func(err error) {
-			log.WithError(err).Error("netpoll: wait loop error")
-		},
-	})
+	poller, err := netpoll.New(nil)
 	if err != nil {
 		return err
 	}
-	defer poller.Close()
 
 	// Get netpoll descriptor with EventRead|EventEdgeTriggered.
 	desc, err := netpoll.HandleListener(l, netpoll.EventRead|netpoll.EventEdgeTriggered)
 	if err != nil {
 		return err
 	}
-	defer desc.Close()
 
 	var (
 		runc = make(chan bool, 1)
